+++ cuppa/cpp/run_patched_boost_test.py
@@ -739,10 +739,11 @@ class RunPatchedBoostTest:
         return line
 
 
-    def __init__( self, expected, final_dir, working_dir=None, **ignored_kwargs ):
+    def __init__( self, expected, final_dir, working_dir=None, inherit_process_env=None, **ignored_kwargs ):
         self._expected = expected
         self._final_dir = final_dir
         self._working_dir = working_dir
+        self._inherit_process_env = inherit_process_env
 
 
     def __call__( self, target, source, env ):
@@ -827,7 +828,8 @@ class RunPatchedBoostTest:
                                                     process_stderr,
                                                     shlex.split( test_command ),
                                                     cwd=working_dir,
-                                                    scons_env=env )
+                                                    scons_env=env,
+                                                    inherit_process_env=self._inherit_process_env )
 
         return return_code, process_stdout.tests()